requirements.txt
asgiref==3.6.0
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.3.0
Django==4.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
pyodbc==4.0.35
pytz==2023.3
requests==2.31.0
sqlparse==0.4.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
whitenoise==6.2.0
